+++ src/Validators/Billet/FineValidator.php
@@ -1,0 +1,97 @@
+<?php
+
+namespace WeDevBr\Bankly\Validators\Billet;
+
+use WeDevBr\Bankly\Types\Billet\Fine;
+
+/**
+ * FineValidator class
+ *
+ * PHP version 8.0|8.1
+ *
+ * @author    WeDev Brasil Team <contato@wedev.software>
+ * @author    Marco Belmont <marco.santos@wedev.software>
+ * @copyright 2021 We Dev Tecnologia Ltda
+ * @link      https://github.com/wedevBr/bankly-laravel/
+ */
+class FineValidator
+{
+    /** @var Fine */
+    private $fine;
+
+    /**
+     * @param Fine $fine
+     */
+    public function __construct(Fine $fine)
+    {
+        $this->fine = $fine;
+    }
+
+    /**
+     * Validate the attributes of the fine class
+     *
+     * @return void
+     */
+    public function validate(): void
+    {
+        $this->validateStartDate();
+        $this->validateValue();
+        $this->validateType();
+    }
+
+    /**
+     * This validates the start date
+     *
+     * @return void
+     * @throws \InvalidArgumentException
+     */
+    private function validateStartDate()
+    {
+        $startDate = $this->fine->startDate;
+        try {
+            $date = now()->createFromFormat('Y-m-d', $startDate);
+            if (!$date->gt(now())) {
+                throw new \InvalidArgumentException('start date must be greater than the current date');
+            }
+        } catch (\Throwable $th) {
+            throw new \InvalidArgumentException('start date should be a valid date');
+        }
+    }
+
+    /**
+     * This validates the value
+     *
+     * @return void
+     * @throws \InvalidArgumentException
+     */
+    private function validateValue()
+    {
+        $value = $this->fine->value;
+        if (empty($value) || !is_string($value) || !is_numeric($value) || $value <= 0) {
+            throw new \InvalidArgumentException('value should be a numeric string and greater than zero');
+        }
+    }
+
+    /**
+     * This validates a type
+     *
+     * @return void
+     * @throws \InvalidArgumentException
+     */
+    private function validateType()
+    {
+        $type = $this->fine->type;
+        if (empty($type) || !is_string($type)) {
+            throw new \InvalidArgumentException('type should be a string');
+        }
+
+        $types = [
+            'FixedAmount',
+            'Percent',
+            'Free',
+        ];
+        if (!in_array($type, $types)) {
+            throw new \InvalidArgumentException('this type is not valid');
+        }
+    }
+}